re_framework: extract the Store trait from the turso impl (#186)
store.rs is now the backend-neutral contract: the row/write/token types,
new_generation(), the Store trait (async_trait — the global store() needs
dyn dispatch), init_store(), and a store::contract module holding the four
unit tests as reusable functions over &dyn Store. turso_store.rs is the
one implementation (TursoStore + SQL + create_tables + init_turso_store),
and its test module just runs the contract suite — a second backend passes
the exact same tests or it's wrong.

Bodies are unchanged; callers are unchanged (store() now returns
&'static dyn Store). The contract semantics every backend must honor —
atomic commit unit, version+generation CAS, generation-guarded ack/poison,
two-tier never-regressing dedup, ordered/paged queries, bidirectional
delete purge — are documented in the README framework section, which also
picks up the two runtime invariants from the second review pass
(sole-writer loop, definitive-verdict dispatch) as prose rather than
source comments.

New dep: async-trait (native async fn in traits is still not
dyn-compatible).
