+++ re_framework/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 [dependencies]
 tokio = { version = "1.36", features = ["full"] }
 anyhow = "1.0"
+async-trait = "0.1"
 chrono = { version = "0.4", features = ["serde"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
