+++ re_framework/src/lib.rs
@@ -3,6 +3,7 @@ mod handle;
 mod machine;
 mod store;
 mod sweep;
+mod turso_store;
 
 #[cfg(test)]
 mod smoke;
@@ -10,5 +11,5 @@ mod smoke;
 pub use effects::Effects;
 pub use handle::{handle, register, StateMachineHandle};
 pub use machine::{EntityId, Identified, Scheduled, StateMachine};
-pub use store::init_turso_store;
+pub use turso_store::init_turso_store;
 pub use sweep::start_sweeper;
